# docs/api-spec.yml
openapi: 3.0.3
info:
  title: JTrack API
  version: 1.0.0
  description: |
    REST API for JTrack CRM (field service, location-scoped multi-tenancy, offline sync).
servers:
  - url: http://localhost:3001
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users in location
      parameters:
        - $ref: '#/components/parameters/XLocationId'
      responses:
        '200':
          description: User list for location
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserWithMembership'
    post:
      tags: [Users]
      summary: Create user (optionally assign to location)
      parameters:
        - $ref: '#/components/parameters/XLocationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserInput'
      responses:
        '200':
          description: Created/upserted user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/invite:
    post:
      tags: [Users]
      summary: Invite user to location
      parameters:
        - $ref: '#/components/parameters/XLocationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserInput'
      responses:
        '200':
          description: Invitation created/updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by id
      parameters:
        - $ref: '#/components/parameters/XLocationId'
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: User or null
          content:
            application/json:
              schema:
                nullable: true
                allOf:
                  - $ref: '#/components/schemas/User'
    patch:
      tags: [Users]
      summary: Update user
      parameters:
        - $ref: '#/components/parameters/XLocationId'
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserInput'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags: [Users]
      summary: Delete user and memberships
      parameters:
        - $ref: '#/components/parameters/XLocationId'
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /tickets:
    get:
      tags: [Tickets]
      summary: List tickets in location
      parameters:
        - $ref: '#/components/parameters/XLocationId'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/TicketStatus'
        - in: query
          name: assignedToUserId
          schema:
            type: string
      responses:
        '200':
          description: Ticket list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
    post:
      tags: [Tickets]
      summary: Create ticket
      parameters:
        - $ref: '#/components/parameters/XLocationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketInput'
      responses:
        '200':
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
  /tickets/{id}:
    get:
      tags: [Tickets]
      summary: Get ticket details
      parameters:
        - $ref: '#/components/parameters/XLocationId'
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetails'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Tickets]
      summary: Update ticket
      parameters:
        - $ref: '#/components/parameters/XLocationId'
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketInput'
      responses:
        '200':
          description: Updated ticket details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetails'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Tickets]
      summary: Soft-delete ticket
      parameters:
        - $ref: '#/components/parameters/XLocationId'
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tickets/{id}/status:
    patch:
      tags: [Tickets]
      summary: Transition ticket status
      parameters:
        - $ref: '#/components/parameters/XLocationId'
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketStatusInput'
      responses:
        '200':
          description: Updated ticket details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetails'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XLocationId:
      in: header
      name: x-location-id
      required: true
      schema:
        type: string
      description: Active location scope for multi-tenant resources.
    PathId:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
      required: [statusCode, message]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]
      required: [ok]
...
    RoleKey:
      type: string
      enum: [Owner, Manager, Technician]
    UserLocationStatus:
      type: string
      enum: [invited, active, suspended]
    TicketStatus:
      type: string
      enum: [New, Scheduled, InProgress, Done, Invoiced, Paid, Canceled]
    AttachmentKind:
      type: string
      enum: [Photo, File]
    PaymentProvider:
      type: string
      enum: [manual, stripe]
    PaymentStatus:
      type: string
      enum: [Pending, Succeeded, Failed, Refunded]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, email, name, isAdmin, createdAt, updatedAt]
    UserWithMembership:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            role:
              $ref: '#/components/schemas/RoleKey'
            membershipStatus:
              $ref: '#/components/schemas/UserLocationStatus'
          required: [role, membershipStatus]
...
    CreateUserInput:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 8
        role:
          $ref: '#/components/schemas/RoleKey'
        locationId:
          type: string
      required: [email, name]
    UpdateUserInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        isAdmin:
          type: boolean
    InviteResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]
        userId:
          type: string
        status:
          type: string
          enum: [invited]
      required: [ok, userId, status]
...
    Ticket:
      type: object
      properties:
        id:
          type: string
        locationId:
          type: string
        createdByUserId:
          type: string
        assignedToUserId:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/TicketStatus'
        scheduledStartAt:
          type: string
          format: date-time
          nullable: true
        scheduledEndAt:
          type: string
          format: date-time
          nullable: true
        priority:
          type: string
          nullable: true
        totalAmountCents:
          type: integer
          nullable: true
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - locationId
        - createdByUserId
        - title
        - status
        - currency
        - createdAt
        - updatedAt
    TicketComment:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        locationId:
          type: string
        authorUserId:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, ticketId, locationId, authorUserId, body, createdAt, updatedAt]
    TicketAttachment:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        locationId:
          type: string
        uploadedByUserId:
          type: string
        kind:
          $ref: '#/components/schemas/AttachmentKind'
        storageKey:
          type: string
        url:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - ticketId
        - locationId
        - uploadedByUserId
        - kind
        - storageKey
        - url
        - mimeType
        - size
        - createdAt
        - updatedAt
    PaymentRecord:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        locationId:
          type: string
        provider:
          $ref: '#/components/schemas/PaymentProvider'
        amountCents:
          type: integer
        currency:
          type: string
        status:
          $ref: '#/components/schemas/PaymentStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - ticketId
        - locationId
        - provider
        - amountCents
        - currency
        - status
        - createdAt
        - updatedAt
    TicketDetails:
      allOf:
        - $ref: '#/components/schemas/Ticket'
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: '#/components/schemas/TicketComment'
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/TicketAttachment'
            paymentRecords:
              type: array
              items:
                $ref: '#/components/schemas/PaymentRecord'
          required: [comments, attachments, paymentRecords]
    CreateTicketInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        assignedToUserId:
          type: string
        scheduledStartAt:
          type: string
          format: date-time
        scheduledEndAt:
          type: string
          format: date-time
        priority:
          type: string
        totalAmountCents:
          type: integer
          minimum: 0
        currency:
          type: string
          default: EUR
      required: [title]
    UpdateTicketInput:
      allOf:
        - $ref: '#/components/schemas/CreateTicketInput'
        - type: object
          properties:
            status:
              $ref: '#/components/schemas/TicketStatus'
    UpdateTicketStatusInput:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/TicketStatus'
      required: [status]
...